bentoML/requirements.txt
bentoml
tensorflow
pillow
scikit-learn
tf2onnx